+++ windows/install-boot-success-task.ps1
@@ -31,8 +31,8 @@ $actionArgs = "-NoProfile -ExecutionPolicy Bypass -File `"$resolvedScript`" -Con
 $userId = [Security.Principal.WindowsIdentity]::GetCurrent().Name
 
 $action = New-ScheduledTaskAction -Execute "powershell.exe" -Argument $actionArgs
-$trigger = New-ScheduledTaskTrigger -AtLogOn
-$principal = New-ScheduledTaskPrincipal -UserId $userId -LogonType Interactive -RunLevel Highest
+$trigger = New-ScheduledTaskTrigger -AtLogOn -User $userId
+$principal = New-ScheduledTaskPrincipal -UserId $userId -LogonType Interactive -RunLevel Limited
 $settings = New-ScheduledTaskSettingsSet -MultipleInstances IgnoreNew -StartWhenAvailable
 
 Register-ScheduledTask `